+++ components/landing/topCollections/TopCollectionsItem.vue
@@ -55,6 +55,7 @@
 <script lang="ts" setup>
 import { RowSeries, TimeRange } from '@/components/series/types'
 import { calculateUsdFromKsm } from '~~/utils/calculation'
+import { CollectionEntityWithVolumes } from './useTopCollections'
 
 const BasicImage = defineAsyncComponent(
   () => import('@/components/shared/view/BasicImage.vue')
@@ -71,7 +72,7 @@ const BasicMoney = defineAsyncComponent(
 const { urlPrefix } = usePrefix()
 const { $store } = useNuxtApp()
 const props = defineProps<{
-  collection: RowSeries
+  collection: CollectionEntityWithVolumes
   index: number
   timeRange?: TimeRange
 }>()
